requirements.txt
PyMuPDF==1.18.14
python-pptx==0.6.19
pytesseract==0.3.8
Pillow==9.0.0